Python/serviceBusMessages/requirements.txt
azure-common==1.1.16
azure-mgmt-containerinstance==1.2.1